Slight refactoring to simplify handler
Moved the getAccount function to a utils-file, removed from the handler.
Importing this from index from files imported from index seemed odd.
This way the handler is only responsible for handling the event.

The handler also was responsible for sending the slackMessages. This was
moved to the SlackMessage class itself, which is now responsible for
sending itself.

Lastly, reverted single-letter variables since I was immediately
confused by them.