feat: Format cloudtrail error logs
Adds logic to the logs-handler to detect (cloudtrail) error logs,
and adds a message formatter that extracts the error code and message
for these logs, only sending those in the message. The header is based
on the error code, but only if all messages in the event are of the
same type.