feat: Allow exclusion of string patterns from Logs
Allows specifying regex patterns to exclude from (cloudtrail) logs.
Matched exclusion strings are filtered from the log events returned to
the messageFormatter. If all events in a log message are excluded, the
handler will return false, and no message will be sent.

This adds a match for drift detection, which is currently not setup
entirely correctly.